refactor: testing and building out new edit closed group capabilities
