+++ app/build.gradle
@@ -18,6 +18,7 @@ plugins {
     id 'kotlin-kapt'
     id 'com.google.dagger.hilt.android'
     id 'com.google.devtools.ksp' version "$kspVersion"
+    id 'app.cash.molecule'
 }
 
 apply plugin: 'com.android.application'
