+++ app/src/main/java/org/thoughtcrime/securesms/conversation/settings/ConversationSettingsActivity.kt
@@ -14,8 +14,7 @@ import org.thoughtcrime.securesms.PassphraseRequiredActionBarActivity
 import org.thoughtcrime.securesms.conversation.v2.ConversationActivityV2
 import org.thoughtcrime.securesms.database.LokiThreadDatabase
 import org.thoughtcrime.securesms.database.ThreadDatabase
-import org.thoughtcrime.securesms.groups.EditClosedGroupActivity
-import org.thoughtcrime.securesms.mms.GlideApp
+import org.thoughtcrime.securesms.groups.EditLegacyClosedGroupActivity
 import org.thoughtcrime.securesms.showSessionDialog
 import javax.inject.Inject
 
@@ -164,9 +163,9 @@ class ConversationSettingsActivity: PassphraseRequiredActionBarActivity(), View.
             v === binding.editGroup -> {
                 val recipient = viewModel.recipient ?: return
                 if (!recipient.isClosedGroupRecipient || !recipient.isLegacyClosedGroupRecipient) return
-                val intent = Intent(this, EditClosedGroupActivity::class.java)
+                val intent = Intent(this, EditLegacyClosedGroupActivity::class.java)
                 val groupID: String = recipient.address.toGroupString()
-                intent.putExtra(EditClosedGroupActivity.groupIDKey, groupID)
+                intent.putExtra(EditLegacyClosedGroupActivity.groupIDKey, groupID)
                 startActivity(intent)
             }
         }
