+++ app/src/main/java/org/thoughtcrime/securesms/groups/EditClosedGroupLoader.kt
@@ -4,13 +4,13 @@ import android.content.Context
 import org.thoughtcrime.securesms.dependencies.DatabaseComponent
 import org.thoughtcrime.securesms.util.AsyncLoader
 
-class EditClosedGroupLoader(context: Context, val groupID: String) : AsyncLoader<EditClosedGroupActivity.GroupMembers>(context) {
+class EditClosedGroupLoader(context: Context, val groupID: String) : AsyncLoader<EditLegacyClosedGroupActivity.GroupMembers>(context) {
 
-    override fun loadInBackground(): EditClosedGroupActivity.GroupMembers {
+    override fun loadInBackground(): EditLegacyClosedGroupActivity.GroupMembers {
         val groupDatabase = DatabaseComponent.get(context).groupDatabase()
         val members = groupDatabase.getGroupMembers(groupID, true)
         val zombieMembers = groupDatabase.getGroupZombieMembers(groupID)
-        return EditClosedGroupActivity.GroupMembers(
+        return EditLegacyClosedGroupActivity.GroupMembers(
                 members.map {
                     it.address.toString()
                 },
