+++ app/src/main/java/org/thoughtcrime/securesms/groups/compose/EditGroupNavGraph.kt
@@ -1,0 +1,8 @@
+package org.thoughtcrime.securesms.groups.compose
+
+import com.ramcosta.composedestinations.annotation.NavGraph
+
+@NavGraph
+annotation class EditGroupNavGraph(
+    val start: Boolean = false
+)
