+++ app/src/main/res/layout/activity_edit_closed_group.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="org.thoughtcrime.securesms.groups.EditClosedGroupActivity">
+    tools:context="org.thoughtcrime.securesms.groups.EditLegacyClosedGroupActivity">
 
     <LinearLayout
         android:id="@+id/mainContentContainer"
