+++ build.gradle
@@ -17,6 +17,7 @@ buildscript {
         classpath "com.google.gms:google-services:$googleServicesVersion"
         classpath "com.squareup:javapoet:1.13.0"
         classpath "com.google.dagger:hilt-android-gradle-plugin:$daggerVersion"
+        classpath 'app.cash.molecule:molecule-gradle-plugin:1.2.1'
         if (project.hasProperty('huawei')) classpath 'com.huawei.agconnect:agcp:1.9.1.300'
     }
 }
